move files common to all DBMS in ./lib dir
