+++ lib/constants.js
@@ -2,7 +2,7 @@
  * @file Program constants
  */
 
-const packageInfo = require('./package.json');
+const packageInfo = require('../package.json');
 
 
 /**
