+++ lib/error-handler.js
@@ -1,6 +1,6 @@
 const cst = require('./constants');
-const db = require('./lib/db-constants');
-const log = require('./lib/log');
+const db = require('./db-constants');
+const log = require('./log');
 
 /**
  * Deduce which information provoked the error and set questions to replace such information
