+++ lib/exportEntities.js
@@ -3,7 +3,8 @@
  */
 
 const fse = require('fs-extra');
-const cst = require('../constants');
+
+const cst = require('./constants');
 
 /**
  * Take the entities object and create a Json file from it
