+++ lib/prompt.js
@@ -6,10 +6,10 @@ const lodash = require('lodash');
 const inquirer = require('inquirer');
 const fse = require('fs-extra');
 
-const log = require('./lib/log');
+const log = require('./log');
 const cst = require('./constants');
-const validation = require('./lib/validation');
-const db = require('./lib/db-commons');
+const validation = require('./validation');
+const db = require('./db-commons');
 
 const toArray = lodash.values;
 const pickProperty = lodash.mapValues;
