+++ lib/validation.js
@@ -4,8 +4,9 @@
 
 // validation framework
 const joi = require('joi');
+
 const db = require('./db-commons');
-const cst = require('../constants');
+const cst = require('./constants');
 
 /**
  * Validate user input against rules
